requirements.txt
langchain
langgraph
pinecone
sentence-transformers
pypdf
flask
uvicorn
python-dotenv
tqdm